server/requirements.txt
fastapi
uvicorn[standard]
websockets
numpy
pillow
pydub
opencv-python
pyautogui
audioop-lts
tqdm
